backend/requirements.txt
google-genai==0.3.0
websockets>=13.0,<15.0
python-dotenv==1.0.1